refactor:#195 api 함수 이동 및 이름 변경
