+++ src/apis/analysis-request.api.ts
@@ -1,4 +1,5 @@
 'use server';
+
 import { camelToSnakeObject, snakeToCamelObject } from '@/utils/camelize.util';
 import { getServerClient } from '@/lib/supabase/server';
 import {
@@ -89,4 +90,18 @@ export const updateCaloriesAnalysisResult = async (food: AiResponseDTO): Promise
   const { error } = await supabase.from('ai_responses').update(camelToSnakeObject(food)).eq('id', id);
 
   return { error };
+};
+
+/**
+ * 분석시 저장한 상세 식사 데이터를 데이터베이스에서 삭제합니다.
+ *
+ * @param {string}  id 로그인한 유저 id
+ * @throws {Error} 데이터베이스 오류 발생 시 에러를 던집니다
+ */
+export const deleteAnalysisData = async () => {
+  const supabase = getServerClient();
+  const { id } = await getUser();
+  const { error: requestsError } = await supabase.from('ai_requests').delete().eq('user_id', id);
+  const { error: requestsDetailError } = await supabase.from('ai_responses').delete().eq('user_id', id);
+  if (requestsDetailError || requestsError) throw new Error('등록된 식사정보 삭제에 실패하였습니다. ');
 };
